Refactor Number 4
