+++ api/Makefile
@@ -1,0 +1,3 @@
+
+all:
+	mvn compile quarkus:dev
