+++ api/src/main/java/org/alliancegenome/curation_api/config/RestApplication.java
@@ -3,7 +3,7 @@ package org.alliancegenome.curation_api.config;
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
 
-import org.alliancegenome.curation_api.auth.Secured;
+import org.alliancegenome.curation_api.model.auth.Secured;
 import org.eclipse.microprofile.openapi.annotations.*;
 import org.eclipse.microprofile.openapi.annotations.enums.SecuritySchemeType;
 import org.eclipse.microprofile.openapi.annotations.info.Info;
