+++ service/src/main/java/org/alliancegenome/curation_api/services/loads/BulkLoadFileHistoryService.java
@@ -1,14 +1,14 @@
 package org.alliancegenome.curation_api.services.loads;
 
 import javax.annotation.PostConstruct;
-import javax.enterprise.context.RequestScoped;
+import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 import org.alliancegenome.curation_api.dao.loads.BulkLoadFileHistoryDAO;
 import org.alliancegenome.curation_api.model.entities.bulkloads.BulkLoadFileHistory;
 import org.alliancegenome.curation_api.services.base.BaseEntityCrudService;
 
-@RequestScoped
+@ApplicationScoped
 public class BulkLoadFileHistoryService extends BaseEntityCrudService<BulkLoadFileHistory, BulkLoadFileHistoryDAO> {
 	
 	@Inject
