+++ service/src/main/java/org/alliancegenome/curation_api/services/ontology/FbdvTermService.java
@@ -1,14 +1,14 @@
 package org.alliancegenome.curation_api.services.ontology;
 
 import javax.annotation.PostConstruct;
-import javax.enterprise.context.RequestScoped;
+import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 import org.alliancegenome.curation_api.dao.ontology.FbdvTermDAO;
 import org.alliancegenome.curation_api.model.entities.ontology.FBDVTerm;
 import org.alliancegenome.curation_api.services.base.BaseOntologyTermService;
 
-@RequestScoped
+@ApplicationScoped
 public class FbdvTermService extends BaseOntologyTermService<FBDVTerm, FbdvTermDAO> {
 
 	@Inject FbdvTermDAO fbdvTermDAO;
