+++ service/src/main/java/org/alliancegenome/curation_api/services/ontology/NcbiTaxonTermService.java
@@ -1,7 +1,7 @@
 package org.alliancegenome.curation_api.services.ontology;
 
 import javax.annotation.PostConstruct;
-import javax.enterprise.context.RequestScoped;
+import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 import org.alliancegenome.curation_api.dao.ontology.NcbiTaxonTermDAO;
@@ -12,7 +12,7 @@ import org.alliancegenome.curation_api.services.base.BaseOntologyTermService;
 import lombok.extern.jbosslog.JBossLog;
 
 @JBossLog
-@RequestScoped
+@ApplicationScoped
 public class NcbiTaxonTermService extends BaseOntologyTermService<NCBITaxonTerm, NcbiTaxonTermDAO> {
 
 	@Inject NcbiTaxonTermDAO ncbiTaxonTermDAO;
